Don't chop levels
@amykhoover Keeping you in the loop.

At Dagstuhl, the quick and dirty solution to being able to bin the original Mario levels was to chop the length to be divisible by the number of level segments that we were evolving. This cuts off part of the original levels, but was required by the old binning schemes I was using. I have now disabled this chopping, since the new binning scheme doesn't care about being divisible by segments. This means we'll need to re-run the Mario stuff to see what changed, but we're not quite ready to do that yet ... need to do some other fixes first.