Changed level rendering
@amykhoover Second set of updates.

These changes fix the way that saved levels from the original game are rendered in the saved PNG files. For evaluation, buffer zones get added at the front and end of every level. This makes sense for evolved levels, since otherwise there might not even be a place for Mario to stand. It makes less sense for actual Mario levels, except that when Vanessa made the new tile encoding for the levels, she apparently removed buffer zones at the starts of actual levels in anticipation of them being added back in by the code, so we need the buffer zones after all. However, the code was originally set up to not render these buffer zones in the saved images, since they are not the evolved part of the level. I've now made it so that the code will render those buffer regions, which is important for levels from the original game. However, I think that evolved levels will still not render the buffer portion, and I'm now no longer sure if that is appropriate. Need to think about the matter a bit more before actually trying to evolve anything.