Don't allow resume with MAP Elites
@amykhoover
For NSGA2 and other generation-based algorithms, my code allowed resuming from a crash. However, my MAP-Elites code does not because it would slow things down too much to maintain the logs in a way that would make this possible. I don't have a solution for this, but the code was pretending it could safely resume. I've changed the code so that it gives the proper error message when the attempt is made.